Add conditional copy for the Conversion Reporting Notification CTA.
